+++ test/test_async.js
@@ -65,7 +65,7 @@ test("state transitions with delays", function() {
 
 //-----------------------------------------------------------------------------
 
-test("state transition fired imediately during onleavestate hook", function() {
+test("state transition fired imediately during onleavestate callback", function() {
 
   var fsm = StateMachine.create({
     initial: 'green',
@@ -74,11 +74,13 @@ test("state transition fired imediately during onleavestate hook", function() {
       { name: 'panic', from: 'yellow', to: 'red',    async: true },
       { name: 'calm',  from: 'red',    to: 'yellow', async: true },
       { name: 'clear', from: 'yellow', to: 'green',  async: true }
-  ]});
-
-  fsm.onleavegreen  = function() { this.transition(); }
-  fsm.onleaveyellow = function() { this.transition(); }
-  fsm.onleavered    = function() { this.transition(); }
+    ],
+    callbacks: {
+      onleavegreen:  function() { this.transition(); },
+      onleaveyellow: function() { this.transition(); },
+      onleavered:    function() { this.transition(); }
+    }
+  });
 
   equals(fsm.current, 'green', "initial state should be green");
 
@@ -91,7 +93,7 @@ test("state transition fired imediately during onleavestate hook", function() {
 
 //-----------------------------------------------------------------------------
 
-test("state transition fired during onleavestate hook with delay", function() {
+test("state transition fired during onleavestate callback with delay", function() {
 
   stop(); // doing async stuff - dont run next qunit test until I call start() below
 
@@ -99,13 +101,15 @@ test("state transition fired during onleavestate hook with delay", function() {
     initial: 'green',
     events: [
       { name: 'panic', from: 'green', to: 'red', async: true }
-  ]});
-
-  fsm.onleavegreen = function() { setTimeout(function() { fsm.transition(); }, 10); }
-  fsm.onenterred   = function() { 
-    equals(fsm.current, 'red', "panic event should transition from green to red");
-    start();
-  }
+    ],
+    callbacks: {
+      onleavegreen: function() { setTimeout(function() { fsm.transition(); }, 10); },
+      onenterred:   function() { 
+        equals(fsm.current, 'red', "panic event should transition from green to red");
+        start();
+      }
+    }
+  });
 
                equals(fsm.current, 'green', "initial state should be green");
   fsm.panic(); equals(fsm.current, 'green', "should still be green because we haven't transitioned yet");
@@ -114,7 +118,7 @@ test("state transition fired during onleavestate hook with delay", function() {
 
 //-----------------------------------------------------------------------------
 
-test("state transition fired during onleavestate hook - with MIXED async and non-async transitions!", function() {
+test("state transition fired during onleavestate callback - with MIXED async and non-async transitions!", function() {
 
   var fsm = StateMachine.create({
     initial: 'green',
@@ -122,9 +126,11 @@ test("state transition fired during onleavestate hook - with MIXED async and non
       { name: 'warn',   from: 'green',           to: 'yellow', async: true }, // leave green     async
       { name: 'panic',  from: 'green',           to: 'red'                 }, // leave green non-async
       { name: 'reset',  from: ['yellow', 'red'], to: 'green'               }
-  ]});
-
-  fsm.onleavegreen  = function() { this.transition(); }
+    ],
+    callbacks: {
+      onleavegreen: function() { this.transition(); }
+    }
+  });
 
                equals(fsm.current, 'green',  "initial state should be green");
   fsm.warn();  equals(fsm.current, 'yellow', "warn  event should transition from green  to yellow");
@@ -154,7 +160,7 @@ test("state transitions using onleavestate run-time return value instead of desi
   fsm.calm();       equals(fsm.current, 'yellow', "calm event should transition from red to yellow");
   fsm.clear();      equals(fsm.current, 'green',  "clear event should transition from yellow to green");
 
-  // but add hooks that return false and it magically becomes asynchronous
+  // but add callbacks that return false and it magically becomes asynchronous
 
   fsm.onleavegreen  = function() { return false; }
   fsm.onleaveyellow = function() { return false; }
@@ -216,7 +222,9 @@ test("state transition fired without completing previous transition", function()
 
 //-----------------------------------------------------------------------------
 
-test("hooks are called when appropriate", function() {
+test("callbacks are ordered correctly", function() {
+
+  var called = [];
 
   var fsm = StateMachine.create({
     initial: 'green',
@@ -225,30 +233,27 @@ test("hooks are called when appropriate", function() {
       { name: 'panic', from: 'yellow', to: 'red',    async: true },
       { name: 'calm',  from: 'red',    to: 'yellow', async: true },
       { name: 'clear', from: 'yellow', to: 'green',  async: true },
-  ]});
+    ],
+    callbacks: {
+      onchangestate: function(event,from,to) { called.push('onchange from ' + from + ' to ' + to); },
 
-  var called = [];
+      onentergreen:  function() { called.push('onentergreen');     },
+      onleavegreen:  function() { called.push('onleavegreen');     },
+      onenteryellow: function() { called.push('onenteryellow');    },
+      onleaveyellow: function() { called.push('onleaveyellow');    },
+      onenterred:    function() { called.push('onenterred');       },
+      onleavered:    function() { called.push('onleavered');       },
 
-  // generic state hook
-  fsm.onchangestate = function(event,from,to) { called.push('onchange from ' + from + ' to ' + to); };
-
-  // state hooks
-  fsm.onentergreen    = function() { called.push('onentergreen');     };
-  fsm.onleavegreen    = function() { called.push('onleavegreen');     };
-  fsm.onenteryellow   = function() { called.push('onenteryellow');    };
-  fsm.onleaveyellow   = function() { called.push('onleaveyellow');    };
-  fsm.onenterred      = function() { called.push('onenterred');       };
-  fsm.onleavered      = function() { called.push('onleavered');       };
-
-  // event hooks
-  fsm.onbeforewarn    = function() { called.push('onbeforewarn');     };
-  fsm.onafterwarn     = function() { called.push('onafterwarn');      };
-  fsm.onbeforepanic   = function() { called.push('onbeforepanic');    };
-  fsm.onafterpanic    = function() { called.push('onafterpanic');     };
-  fsm.onbeforecalm    = function() { called.push('onbeforecalm');     };
-  fsm.onaftercalm     = function() { called.push('onaftercalm');      };
-  fsm.onbeforeclear   = function() { called.push('onbeforeclear');    };
-  fsm.onafterclear    = function() { called.push('onafterclear');     };
+      onbeforewarn:  function() { called.push('onbeforewarn');     },
+      onafterwarn:   function() { called.push('onafterwarn');      },
+      onbeforepanic: function() { called.push('onbeforepanic');    },
+      onafterpanic:  function() { called.push('onafterpanic');     },
+      onbeforecalm:  function() { called.push('onbeforecalm');     },
+      onaftercalm:   function() { called.push('onaftercalm');      },
+      onbeforeclear: function() { called.push('onbeforeclear');    },
+      onafterclear:  function() { called.push('onafterclear');     }
+    }
+  });
 
   called = [];
   fsm.warn();       deepEqual(called, ['onbeforewarn', 'onleavegreen']);